requirements.txt
afastapi
uvicorn[standard]
pdfplumber
python-multipart